pyproject.toml
[project]
name = "rag-doc-assistant"
version = "1.0.0"
description = "RAG based documentation assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.1",
    "pdfminer.six>=20240706",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.4",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "python-magic-bin>=0.4.14",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "tqdm>=4.67.1",
]
